refactor: how we handle global errors
