+++ src/legacy/auth/v3/context.ts
@@ -1,6 +1,6 @@
 import { TConnectContextRequest, TCallbackContextRequest } from './types'
 import { NextFunction, Response } from 'express'
-import { UserError } from '../../../errors'
+import { UserError } from '../../errors'
 import { NoAuthInProgress } from './errors'
 import { asyncMiddleware } from '../../errorHandler'
